beth_health_journey_app/requirements.txt
python-dotenv==1.0.0
notion-client==2.2.1
supabase==2.0.3
openai==1.12.0
pandas==2.1.3
matplotlib==3.8.1
seaborn==0.13.0
PyPDF2==3.0.1
pytesseract==0.3.10
scikit-learn==1.3.2
flask==2.3.3
pytest==7.4.3
black==23.11.0
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0 